19-B Corrigiendo el tipo de integer para best_answer_id
ns estancamos aqui.... cuando corremos aamf nos da error por la referencia de best_answer_id
fortunately I got fixed. the solution was to change *best_answer_id type from integer to BigInteger

+++ database/migrations/2020_11_13_070816_create_questions_table.php
@@ -26,7 +26,7 @@ class CreateQuestionsTable extends Migration
             $table->unsignedInteger('answers')->default(0);
             # how many people are voting the question
             $table->integer('votes')->default(0);
-            $table->unsignedInteger('best_answer_id')->nullable();
+            $table->unsignedBigInteger('best_answer_id')->nullable();
             $table->unsignedBigInteger('user_id');
 
             $table->timestamps();
